--- src/jack_flags/mod.rs
@@ -1,14 +1,0 @@
-/// Contains `ClientOptions` flags used when opening a client.
-pub mod client_options;
-
-/// Contains `ClientStatus` flags which describe the status of a
-/// Client.
-pub mod client_status;
-
-/// Contains `PortFlags` flags which can be used when implementing the
-/// `PortSpec` trait.
-pub mod port_flags;
-
-pub use client_options::ClientOptions;
-pub use client_status::ClientStatus;
-pub use port_flags::PortFlags;
