+++ src/port/mod.rs
@@ -1,5 +1,9 @@
 mod audio;
 mod midi;
+mod port;
+pub mod port_flags;
 
+pub use self::port::*;
 pub use self::audio::{AudioInSpec, AudioInPort, AudioOutSpec, AudioOutPort};
 pub use self::midi::{MidiInSpec, MidiInPort, MidiIter, MidiOutSpec, MidiOutPort, RawMidi};
+pub use self::port_flags::PortFlags;
