+++ src/test/test_client_port.rs
@@ -1,4 +1,5 @@
-use super::super::*;
+use prelude::*;
+use jack_utils::*;
 
 fn open_test_client(name: &str) -> Client {
     default_sleep();
